added remove comments
