+++ pubky-homeserver/src/persistence/files/lmdb_to_opendal_migrator.rs
@@ -1,3 +1,7 @@
+//!
+//! TODO: Remove this whole module after the file migration is complete.
+//! 
+
 use super::{FileIoError, FileService};
 use crate::persistence::lmdb::{
     tables::files::{Entry, FileLocation},
