+++ pubky-homeserver/src/persistence/files/mod.rs
@@ -5,7 +5,7 @@ mod file_service;
 mod file_stream_type;
 mod lmdb_to_opendal_migrator;
 mod opendal_service;
-mod write_disk_quota_enforcer;
+
 
 pub use file_io_error::{FileIoError, WriteStreamError};
 pub(crate) use file_metadata::{FileMetadata, FileMetadataBuilder};
@@ -13,4 +13,3 @@ pub use file_service::FileService;
 pub use file_stream_type::FileStream;
 pub use lmdb_to_opendal_migrator::LmDbToOpendalMigrator;
 pub use opendal_service::OpendalService;
-pub use write_disk_quota_enforcer::is_size_hint_exceeding_quota;
